requirements.txt
flask
flask-smorest
flask-sqlalchemy
flask-jwt-extended
flask-migrate
flask-cors
passlib
python-dotenv



